refactor(db): split drizzle schema into auth + multi-tenant contexts
BetterAuth-owned tables (user/session/account/verification/twoFactor/passkey)
stay in schema/auth.ts; org/member/invitation extracted to schema/multi-tenant.ts.
Combined schema namespace re-exported from src/index.ts (outside drizzle-kit's
schema glob to avoid duplicate-table errors) preserves zero call-site change.

RGPD fields (pendingDeletionUntil/deletedAt/lastExportRequestedAt) remain on
user (BetterAuth-owned column extensions). webhooks.ts updated to import
organization from ./multi-tenant.

Closes ROADMAP 0.1 (Phase 0 foundation closeout).
