+++ ROADMAP.md
@@ -31,7 +31,7 @@ Each phase assumes the previous done. Items inside a phase parallelizable. Order
 | # | Item | Status |
 |---|---|---|
 | 0.0 | Clone-ability bootstrap | **5/7 done** — remaining: CI smoke test + `docs/QUICKSTART.md` |
-| 0.1 | DB schema split per context (currently `auth.ts` bundles auth + multi-tenant + RGPD) | TODO — before Billing |
+| 0.1 | DB schema split per context (`auth.ts` → `auth.ts` + `multi-tenant.ts`) | DONE — BetterAuth-owned tables stay in `auth.ts`; `organization`/`member`/`invitation` extracted to `multi-tenant.ts`. RGPD = 3 cols on `user` (BetterAuth-owned), inline. Combined `schema` re-export preserves call sites (zero refactor). |
 | 0.2 | Health probes `/livez` + `/readyz` + `/startupz` (K8s 2026, IETF `draft-inadarei`, registry pattern) | TODO — graceful shutdown wired to `/readyz` for zero-downtime deploys |
 | 0.3 | Backups `pg_dump` + restore tested (3-2-1 rule, 30d retention, monthly automated restore-test) | TODO — RPO/RTO in `docs/DISASTER-RECOVERY.md` |
 | 0.4 | Sentry + OpenTelemetry + Prometheus `/metrics` (3 detachable ports + NoOp default) | TODO — subscribers now trivial via `onEvent` (event-driven foundation) |
@@ -881,7 +881,7 @@ The **Billing** section above lays the foundation: `PLANS` config, `useEntitleme
 - [x] **Back step 1 — pivot `apps/api/src/` to `modules/<context>/`**: `modules/uploads/`, `modules/rgpd/` shipped (auth + organizations stay at `apps/api/src/auth.ts` — BetterAuth singleton, no DDD). Each contains `application/{dto,services,ports for module-private interfaces}` + `infrastructure/{services,repositories}` + `routes.ts` (where applicable) + `module.ts`. `adapters/` removed; `infrastructure/` is the DDD-canonical naming. **`modules/email/` deliberately not created** — email is pure infra (no domain, no use cases, no routes), lives in `shared/services/email.service.ts` consuming `shared/ports/email.port.ts` (rule "shared kernel" in CLAUDE.md).
 - [x] **Back step 2 — extract `shared/`**: `apps/api/src/shared/{middleware,ports,services}/` + process-level singletons at the root. Cross-context port interfaces (`IStorageService`, `IEmailService`) live in `shared/ports/`; cross-context impls (`ResendEmailService`) in `shared/services/`. The composed env-driven gate for `/internal/*` lives in `shared/internal-routes/internal-layers.ts` (single source — any future module exposing `/internal/*` consumes the same `internalLayers`); the whole `/internal/*` concern (signature primitives + middlewares + signed-fetch client) is grouped under `shared/internal-routes/`. `common/` deleted.
 - [x] **Back step 3 — `module.ts` per context (split form)**: each `module.ts` defines an inwire `defineModule()` (typed prerequisites local to the module). `di/container.ts` chains `.addModule(emailModule).addModule(uploadsModule).addModule(rgpdModule)`. Routes stay in `routes.ts` / `internal.routes.ts` and are imported by `index.ts` directly — splitting DI from routes avoids a `module.ts → routes.ts → di/container.ts → module.ts` cycle (Biome `noImportCycles` flagged it; the SOTA is to keep DI wiring and route mounting on separate import graphs).
-- [ ] **Back step 4 — split DB schema**: today `packages/drizzle/src/schema/auth.ts` is the single file (auth + multi-tenant + rgpd fields all bundled). Split into `packages/drizzle/src/schema/<context>.ts` (auth, organization, uploads when keys land in DB, rgpd, …), barrel `index.ts` re-exports. Each module owns its tables. Removing a module = remove the `export *` line + revert the migration. Last front of the vertical-slice migration; ship before Phase B (Billing) so the new `subscription` / `payment` tables land in their own file from day one.
+- [x] **Back step 4 — split DB schema**: `packages/drizzle/src/schema/auth.ts` → `auth.ts` (BetterAuth-owned: user/session/account/verification/twoFactor/passkey) + `multi-tenant.ts` (organization/member/invitation). RGPD fields stay on `user` (BetterAuth-owned column extensions). Combined `schema` namespace re-export preserved in `src/index.ts` → zero call-site change. New billing tables will land in `packages/drizzle/src/schema/billing.ts` from day one.
 - [ ] **Removability dry-run** on the smallest module (probably `rgpd` since fresh in memory, or `uploads` if smaller) — delete it end-to-end, run `pnpm ci:check`, document the diff in `docs/HISTORY.md` as the canonical "how to remove a feature" example.
 - [ ] **Removability CI gate (phase 2 — deferred until pattern stabilizes)**: script `scripts/check-removability.ts` that picks a random module, snapshots, removes, type-checks, restores. Optional weekly cron in CI; promote to PR-blocking once stable.
